Java side of lifted client, plus some review feedback applied.
