+++ runtime/src/main/scala/akka/grpc/RequestBuilder.scala
@@ -1,0 +1,31 @@
+/**
+ * Copyright (C) 2009-2018 Lightbend Inc. <https://www.lightbend.com>
+ */
+package akka.grpc
+
+import akka.annotation.DoNotInherit
+
+import scala.concurrent.duration.FiniteDuration
+
+/**
+ * Request builder for requests providing per call specific metadata capabilities in
+ * addition to the client instance default options provided to it through [[GrpcClientSettings]] upon creation.
+ *
+ * Instances are immutable so can be shared and re-used but are backed by the client that created the instance,
+ * so if that is stopped the invocations will fail.
+ *
+ * Not for user extension
+ */
+@DoNotInherit
+trait RequestBuilder[Req, Res] {
+
+  /** FIXME docs */
+  def addMetadata(key: String, value: String): RequestBuilder[Req, Res]
+  /** FIXME docs */
+  def withDeadline(deadline: FiniteDuration): RequestBuilder[Req, Res]
+
+  /**
+   * Invoke the gRPC method with the additional metadata added
+   */
+  def invoke(request: Req): Res
+}
